+++ lib/mix/tasks/se_es_clean.ex
@@ -1,15 +1,16 @@
 defmodule Mix.Tasks.Se.EsClean do
-  use Mix.Task
-  alias TdSe.ESClientApi
-
   @moduledoc """
-    Run elastic clean
+  Run elastic clean
   """
+
+  use Mix.Task
+
+  alias TdSe.ESClientApi
 
   @shortdoc "Create test mappings in elasticsearch"
   def run(__args) do
-    Mix.Task.run "app.start"
-    ESClientApi.delete_indexes
-    ESClientApi.delete_aliases
+    Mix.Task.run("app.start")
+    ESClientApi.delete_indexes()
+    ESClientApi.delete_aliases()
   end
 end