+++ rel/bin/start.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+
+set -o errexit
+set -o xtrace
+
+bin/td_se start