+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Changed
 
-- [TD-2272] Filter deleted structures on search
+- [TD-2271] Filter deleted structures on search
 
 ## [3.8.0] 2019-10-14